Continued table cleanup
